intermediate commit for TG

+++ src/pothen_fan.cpp
@@ -9,7 +9,6 @@
 #include "graphtypes.h"
 #include "pothen_fan.h"
 
-
 inline bool lookahead_step_atomic(
 		const Vertex x0,
 		const Graph& g,const Vertex first_right, VertexVector& mate,
@@ -148,7 +147,7 @@ void parallel_pothen_fan(const Graph& g, Vertex first_right, VertexVector& mate,
 		std::vector<PathElement> stack;
 #pragma omp parallel num_threads(nt) private(stack)
 #pragma omp for
-		for (size_t i = 0; i < unmatched_size; i++) {
+		for (int i = 0; i < unmatched_size; i++) {
             Vertex v = unmatched[i];
 		//for (int v = 0; v < first_right; v++) {
 
@@ -459,10 +458,6 @@ void pothen_fan(const Graph& g, const Vertex first_right, VertexVector& mate) {
 	delete[] visited;
     delete[] lookahead;
 }
-
-
-
-
 
 bool find_path(const Vertex x0, const Graph& g,const Vertex first_right, VertexVector& mate, bool* visited, std::vector<FindPathElement>& stack) {
